requirements.txt
fitz==0.0.1.dev2
pandas==2.1.4
pdfminer.six==20221105
pypdf==3.17.4
PyPDF2==3.0.1
tqdm==4.66.1
